+++ testsuite/enum_nthash.cpp
@@ -311,19 +311,19 @@ static void enumerate_nt_hash_template(const char *inputfilename,
     throw msg;
   }
   delete hashed_qgram_packer;
-  printf("# num_of_sequences\t%lu\n",seqnum);
-  printf("# total_length\t%lu\n",total_length);
-  printf("# max_sequence_length\t%lu\n",max_sequence_length);
+  printf("# num_of_sequences\t%zu\n",seqnum);
+  printf("# total_length\t%zu\n",total_length);
+  printf("# max_sequence_length\t%zu\n",max_sequence_length);
   printf("# num_of_sequences.bits\t%d\n",gttl_required_bits(seqnum));
   printf("# max_sequence_length.bits\t%d\n",
           gttl_required_bits<size_t>(max_sequence_length));
-  printf("# count_all_qgrams\t%lu\n",count_all_qgrams);
-  printf("# sum_hash_values\t%" PRIu64 "\n",sum_hash_values);
-  printf("# bytes_unit_sum\t%zu\n",bytes_unit_sum);
+  printf("# count_all_qgrams\t%zu\n",count_all_qgrams);
+  printf("# sum_hash_values\t%zu\n",(size_t) sum_hash_values);
+  printf("# bytes_unit_sum\t%zu\n", (size_t) bytes_unit_sum);
   if constexpr (with_rc)
   {
-    printf("# sum_rc_hash_values\t%" PRIu64 "\n", sum_rc_hash_values);
-    printf("# bytes_unit_sum_rc\t%zu\n",bytes_unit_sum_rc);
+    printf("# sum_rc_hash_values\t%zu\n",(size_t) sum_rc_hash_values);
+    printf("# bytes_unit_sum_rc\t%zu\n", (size_t) bytes_unit_sum_rc);
   }
   gttl_fp_type_close(in_fp);
 }
